requirements.txt
easyocr
nltk
numpy
streamlit
streamlit-authenticator==0.1.5
Pillow

